Add cast to deal with max() requiring identical types on 32-bit systems

+++ sst/elements/memHierarchy/memNIC.cc
@@ -38,7 +38,7 @@ int MemNIC::addrForDest(const std::string &target)
 int MemNIC::getFlitSize(MemEvent *ev)
 {
     /* addr (8B) + cmd (1B) + size */
-    return std::max((9 + ev->getSize()) / flitSize, 1UL);
+    return std::max((unsigned long)((9 + ev->getSize()) / flitSize), 1UL);
 }
